requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
